Centralize publish docs preparation for all public packages
