+++ .gitignore
@@ -3,6 +3,9 @@ dist/
 .flue-vite/
 .flue-vite.wrangler.jsonc
 
+# Generated by scripts/prepare-publish.mjs at publish time.
+packages/*/docs/
+
 plans/
 
 # caches
