+++ packages/cli/.gitignore
@@ -1,5 +1,2 @@
 # Generated by scripts/generate-connector-index.ts at prebuild time.
 bin/_connectors.generated.ts
-
-# Generated by scripts/prepare-publish.mjs at publish time.
-docs/
