+++ packages/postgres/package.json
@@ -18,7 +18,8 @@
 	"main": "./dist/index.mjs",
 	"types": "./dist/index.d.mts",
 	"files": [
-		"dist"
+		"dist",
+		"docs"
 	],
 	"engines": {
 		"node": ">=22.18.0"
