+++ packages/sdk/package.json
@@ -46,7 +46,8 @@
 	"main": "./dist/index.mjs",
 	"types": "./dist/index.d.mts",
 	"files": [
-		"dist"
+		"dist",
+		"docs"
 	],
 	"engines": {
 		"node": ">=22.18.0"
@@ -54,8 +55,7 @@
 	"scripts": {
 		"build": "tsdown",
 		"check:types": "tsc --noEmit",
-		"test": "vitest run",
-		"prepublishOnly": "cp ../../README.md ."
+		"test": "vitest run"
 	},
 	"devDependencies": {
 		"tsdown": "^0.22.0",
